program/requirements.txt
aiohttp==3.8.1
aiosignal==1.3.1
async-timeout==4.0.2
atomicwrites==1.4.1
attrs==22.1.0
base58==2.1.1
bitarray==2.6.0
blinker==1.8.2
cachetools==5.3.3
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cycler==0.12.1
cytoolz==0.11.2
dateparser==1.0.0
distlib==0.3.6
dydx-v3-python==1.9.1
ecdsa==0.16.0
entrypoints==0.4
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.1
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
filelock==3.8.2
fonttools==4.51.0
frozenlist==1.3.3
hexbytes==0.3.0
idna==3.4
importlib-metadata==0.23
ipfshttpclient==0.8.0a2
jsonschema==4.17.3
kiwisolver==1.4.5
lru-dict==1.1.8
lxml==5.2.1
Markdown==3.6
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==9.0.0
mpmath==1.0.0
multiaddr==0.0.9
multidict==6.0.3
mypy-extensions==0.4.3
netaddr==0.8.0
numpy==1.23.5
packaging==22.0
pandas==1.5.2
parsimonious==0.8.1
pathspec==0.10.3
patsy==0.5.3
pillow==10.3.0
platformdirs==2.6.0
pluggy==0.13.1
prometheus_client==0.20.0
protobuf==3.19.5
py==1.11.0
pyarrow==16.0.0
pybit==5.7.0
pycodestyle==2.10.0
pycryptodome==3.16.0
Pygments==2.18.0
PyJWT==2.6.0
pymdown-extensions==10.8.1
pyparsing==3.1.2
pyrsistent==0.19.2
pytest==4.6.11
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
pywin32==306
PyYAML==6.0.1
referencing==0.35.1
regex==2022.10.31
requests==2.28.1
requests-mock==1.6.0
rlp==2.0.1
rpds-py==0.18.1
scipy==1.9.3
six==1.16.0
smmap==5.0.1
soupsieve==2.5
statsmodels==0.13.5
sympy==1.6
tenacity==8.3.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.4
tox==3.13.2
typing_extensions==4.11.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
validators==0.28.1
varint==1.0.2
virtualenv==20.17.1
watchdog==4.0.0
wcwidth==0.2.5
web3==5.31.3
websocket-client==1.8.0
websockets==9.1
yarl==1.8.2
zipp==3.11.0
